requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==24.3.0
beartype==0.19.0
bitrix24-rest==2.0.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
frozenlist==1.5.0
h11==0.14.0
icontract==2.7.1
idna==3.10
magic-filter==1.0.12
more-itertools==10.5.0
multidict==6.1.0
multidimensional_urlencode==0.0.4
numpy==2.2.0
pandas==2.2.3
phonenumbers==8.13.52
propcache==0.2.1
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
yarl==1.18.3
